Commit from Main sustitute branch to Main(Orgin) branch

+++ src/main/java/com/movieplatform/Entity/User.java
@@ -32,7 +32,7 @@ public class User {
     private String gmail;
 
     @Column(name = "admin")
-    private Byte admin;
+    private Integer admin;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "wishlist_id")
